Add ESLint as the default npm test for now
We have a CI in place that will run `npm test` on pull request, so we
might as well use it for something meaningful like ESLint.

The `example/` are ignored for now because of too many errors. If anyone
wants to pick that up, feel free.

This also fixes small ESLint complaints.